requirements.txt
pandas
numpy
matplotlib
pytorch
sklearn
keras==2.2.4
seaborn
lightgbm==2.2.1
xgboost==0.90